+++ tests/stubs/glut/GLUT/glut.h
@@ -1,0 +1,8 @@
+/* Stub for the macOS <GLUT/glut.h> framework header. See ../GL/glut.h. */
+
+#ifndef NANOLANG_TEST_STUB_GLUT_FRAMEWORK_H
+#define NANOLANG_TEST_STUB_GLUT_FRAMEWORK_H
+
+#include "../GL/glut.h"
+
+#endif /* NANOLANG_TEST_STUB_GLUT_FRAMEWORK_H */